1.参数校验错误信息提供抽象方法 2.升级jar依赖版本 3.1.5～

+++ src/main/java/com/jy/medusa/gaze/stuff/Pager.java
@@ -92,8 +92,8 @@ public class Pager<T> implements Serializable {
 		return this;
 	}
 
-	public Long getStartRecord() {
-		return (pageNumber - 1L) * pageSize;
+	public Integer getStartRecord() {
+		return (pageNumber - 1) * pageSize;
 	}
 
 	public List<T> getList() {
